chore: fix infinite loading
