--- packages/modules/data-widgets/src/themesource/datawidgets/web/_datagrid-scroll.scss
@@ -1,27 +1,0 @@
-.container-stuck + .table .th {
-    position: -webkit-sticky; /* Safari */
-    position: sticky;
-    z-index: 50;
-}
-
-/**
- * Another hack for sticky header and filters.
- * Don't use this code anywhere else, and if you can, please remove it.
- * If we moved headers out of the grid body, please, remove this piece of guano.
- */
-.container-stuck + .table .th:has(.filter-selectors.visible) {
-    z-index: 51;
-}
-
-.widget-datagrid-content.infinite-loading {
-    overflow-y: auto;
-    margin-bottom: 20px;
-}
-
-.table {
-    .table-content {
-        &.infinite-loading {
-            overflow-y: scroll;
-        }
-    }
-}
