+++ packages/modules/data-widgets/src/themesource/datawidgets/web/main.scss
@@ -2,7 +2,6 @@
 @import "datagrid";
 @import "datagrid-filters";
 @import "datagrid-design-properties";
-@import "datagrid-scroll";
 @import "drop-down-sort";
 @import "gallery";
 @import "gallery-design-properties";
